# swaggerui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Package main",
    "version": "0.0.1"
  },
  "host": "localhost:10000",
  "basePath": "/",
  "paths": {
    "/api/profile": {
      "get": {
        "description": "Returns a list of all Profiles",
        "tags": [
          "Profile"
        ],
        "operationId": "GetAllProfilesEndpoint",
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      },
      "post": {
        "description": "Creates a Profile in the database",
        "tags": [
          "Profile"
        ],
        "operationId": "CreateProfileEndpoint",
        "parameters": [
          {
            "description": "Profile details",
            "name": "ProfileDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/group": {
      "get": {
        "description": "Returns a list of all ProfileGroups",
        "tags": [
          "ProfileGroup"
        ],
        "operationId": "GetAllProfileGroupsEndpoint",
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      },
      "post": {
        "description": "Creates a ProfileGroup in the database",
        "tags": [
          "ProfileGroup"
        ],
        "operationId": "CreateProfileGroupEndpoint",
        "parameters": [
          {
            "description": "ProfileGroup details",
            "name": "ProfileGroupDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProfileGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/group/{GroupID}": {
      "get": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Returns the ProfileGroup with GroupID {GroupID}.",
        "operationId": "GetProfileGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Updates and returns the ProfileGroup with GroupID {GroupID}.",
        "operationId": "UpdateProfileGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to update",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "New name to update ProfileGroup with",
            "name": "Name",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProfileGroupRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Deletes and returns the ProfileGroup with GroupID {GroupID}.",
        "operationId": "RemoveProfileGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to retrieve",
            "name": "GroupID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/group/{GroupID}/add": {
      "post": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Adds Profiles to the ProfileGroup with GroupID {GroupID} and returns the updated ProfileGroup.",
        "operationId": "AddProfilesToGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to add to",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "Profile IDs to add to the ProfileGroup",
            "name": "ProfileIDs",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/ProfileIDList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/group/{GroupID}/clone": {
      "post": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Clones the ProfileGroup with GroupID {GroupID} and returns the clone.",
        "operationId": "CloneProfileGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to clone",
            "name": "GroupID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/group/{GroupID}/remove": {
      "post": {
        "tags": [
          "ProfileGroup"
        ],
        "summary": "Removes Profiles from the ProfileGroup with GroupID {GroupID} and returns the updated ProfileGroup.",
        "operationId": "RemoveProfilesFromGroupEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of ProfileGroup to add to",
            "name": "GroupID",
            "in": "path"
          },
          {
            "description": "Profile IDs to add to the ProfileGroup",
            "name": "ProfileIDs",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/ProfileIDList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProfileGroup response",
            "schema": {
              "$ref": "#/responses/ProfileGroupResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/{ID}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Returns the Profile with ID {ID}.",
        "operationId": "GetProfileEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Profile to retrieve",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Updates and returns the Profile with ID {ID}.",
        "operationId": "UpdateProfileEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Profile to update",
            "name": "ID",
            "in": "path"
          },
          {
            "description": "Details to update Profile with",
            "name": "ProfileDetails",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/models/CreateOrUpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Deletes and returns the Profile with ID {ID}.",
        "operationId": "RemoveProfileEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Profile to retrieve",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      }
    },
    "/api/profile/{ID}/clone": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "Clones the Profile with ID {ID} and returns the clone.",
        "operationId": "CloneProfileEndpoint",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Profile to clone",
            "name": "ID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile response",
            "schema": {
              "$ref": "#/responses/ProfileResponseSwagger"
            }
          }
        }
      }
    },
...
  }
}